feat: change LanguageSettings component to typescript
